+++ packages/member-base-nestjs-module/src/index.ts
@@ -25,14 +25,10 @@ export * from './decorators/is-public.decorator';
 export * from './decorators/member-id.decorator';
 export * from './decorators/account.decorator';
 export * from './decorators/authenticated.decorator';
+export * from './decorators/has-permission.decorator';
 
 // Helpers
 export * from './helpers/graphql-context-token-resolver';
-
-// GraphQL
-export * from './dto/graphql/token-pair.dto';
-export * from './dto/graphql/base-member.dto';
-export * from './dto/graphql/member-login-log.dto';
 
 // Errors
 export * from './constants/errors/index';