+++ packages/member-base-nestjs-module/src/utils/get-token-from-request.ts
@@ -1,0 +1,13 @@
+import { Request } from 'express';
+
+export const getTokenFromRequest = async (
+  req: Request,
+  cookieMode = false,
+): Promise<string | null> => {
+  return (
+    (cookieMode
+      ? req.cookies.token
+      : (req.headers.authorization ?? '').replace(/^Bearer\s/, '').trim()) ??
+    null
+  );
+};