+++ packages/cms-base-nestjs-graphql-module/CHANGELOG.md
@@ -3,6 +3,34 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.9](https://github.com/Rytass/Utils/compare/@rytass/cms-base-nestjs-graphql-module@0.1.8...@rytass/cms-base-nestjs-graphql-module@0.1.9) (2025-08-04)
+
+**Note:** Version bump only for package @rytass/cms-base-nestjs-graphql-module
+
+## [0.1.8](https://github.com/Rytass/Utils/compare/@rytass/cms-base-nestjs-graphql-module@0.1.7...@rytass/cms-base-nestjs-graphql-module@0.1.8) (2025-08-04)
+
+**Note:** Version bump only for package @rytass/cms-base-nestjs-graphql-module
+
+## [0.1.7](https://github.com/Rytass/Utils/compare/@rytass/cms-base-nestjs-graphql-module@0.1.6...@rytass/cms-base-nestjs-graphql-module@0.1.7) (2025-08-04)
+
+**Note:** Version bump only for package @rytass/cms-base-nestjs-graphql-module
+
+## [0.1.6](https://github.com/Rytass/Utils/compare/@rytass/cms-base-nestjs-graphql-module@0.1.5...@rytass/cms-base-nestjs-graphql-module@0.1.6) (2025-08-04)
+
+**Note:** Version bump only for package @rytass/cms-base-nestjs-graphql-module
+
+## [0.1.5](https://github.com/Rytass/Utils/compare/@rytass/cms-base-nestjs-graphql-module@0.1.4...@rytass/cms-base-nestjs-graphql-module@0.1.5) (2025-08-04)
+
+**Note:** Version bump only for package @rytass/cms-base-nestjs-graphql-module
+
+## [0.1.4](https://github.com/Rytass/Utils/compare/@rytass/cms-base-nestjs-graphql-module@0.1.3...@rytass/cms-base-nestjs-graphql-module@0.1.4) (2025-08-04)
+
+**Note:** Version bump only for package @rytass/cms-base-nestjs-graphql-module
+
+## [0.1.3](https://github.com/Rytass/Utils/compare/@rytass/cms-base-nestjs-graphql-module@0.1.2...@rytass/cms-base-nestjs-graphql-module@0.1.3) (2025-08-04)
+
+**Note:** Version bump only for package @rytass/cms-base-nestjs-graphql-module
+
 ## [0.1.2](https://github.com/Rytass/Utils/compare/@rytass/cms-base-nestjs-graphql-module@0.1.1...@rytass/cms-base-nestjs-graphql-module@0.1.2) (2025-08-04)
 
 **Note:** Version bump only for package @rytass/cms-base-nestjs-graphql-module